backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
bcrypt
pyjwt
python-socketio[asyncio]
email-validator
